Updating all of the projects I've worked on
Most of these are for class, but the snake and other tkinter projects were made for fun. I followed a snake tutorial to create the base game, but I saw area for improvement. It isn't complete but I want to add a start lobby and pause menu. I also added snake phasing because I didn't like the game ending when the snake hit the wall.